communication: rename firmware_update to communication
